Implemented Delim::compressTokens(). Needs testing!
Attempted bugfix for Manager::run() in which Tokens were improperly
separated (e.g. a token like {ls -a} would become separated into {ls},
{-a}), and then were erroneously passed into Manager::evaluate().

Running Delim::compressTokens() should take tokens like {ls} and {-a}
and intelligently splice them together, resulting in the recombined {ls
-a} token. These may then be passed into Manager::evaluate() as normal.

PLEASE read the documentation for Delim::compressTokens() in delim.h.
This is a function that is very specific, and it is very easy to
improperly use this function and break the program!